+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Add badge to event card on index when is sold out
+
 ## [0.0.12] - 2024-01-13
 
 ### Added