+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.15] - 2024-01-22
+
+- Add Payku integration
+
 ## [0.0.14] - 2024-01-22
 
 - Add badge to event card on index when is sold out
@@ -84,7 +88,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Several UI fixes in next event, and add price format
 
-[unreleased]: https://github.com/kbzaso/5lc-sveltkit-sanity/compare/0.0.14...HEAD
+[unreleased]: https://github.com/kbzaso/5lc-sveltkit-sanity/compare/0.0.15...HEAD
+[0.0.14]: https://github.com/kbzaso/5lc-sveltkit-sanity/compare/0.0.14...0.0.15
 [0.0.13]: https://github.com/kbzaso/5lc-sveltkit-sanity/compare/0.0.13...0.0.14
 [0.0.12]: https://github.com/kbzaso/5lc-sveltkit-sanity/compare/0.0.12...0.0.13
 [0.0.11]: https://github.com/kbzaso/5lc-sveltkit-sanity/compare/0.0.11...0.0.12